requirements.txt
pygame>=2.0.0
python-rtmidi>=1.4.0